dupe widgets code gen issue widget_metadata.dart
I was having syntax errors in the generated default_registrar.g.dart file, and the reason was that there were dupes for all the widgets.  The "void withAlign() {" method was repeated over and over as well as all the others. It seemed to add another copy each time the generator ran as if something was "building up".

I took a stab at deduping the widget list and I am not seeing the issue anymore.

I really have no idea out code gen stuff works so not idea if this is actually a good solution or what the root cause is.